factorize out image func in to image_utils
