+++ gee/image_utils.py
@@ -1,0 +1,59 @@
+"""
+Modify, and slice up tif and png images using Python Image Library
+"""
+
+from PIL import Image
+
+
+def combine_tif(input_filebase, bands=["B4","B3","B2"]):
+    """
+    Read tif files in "I" mode - one per specified band, and rescale and combine
+    pixel values to r,g,b values betweek 0 and 255 in a combined output image.
+    Currently assumes that we have three bands.  Need to figure out how to
+    deal with more or fewer...
+    """
+    if len(bands) >= 3:
+        band_dict = {"r": {"band": bands[0],
+                           "min_val": sys.maxsize,
+                           "max_val": -1*sys.maxsize,
+                           "pix_vals": []},
+                     "g": {"band": bands[1],
+                           "min_val": sys.maxsize,
+                           "max_val": -1*sys.maxsize,
+                           "pix_vals": []},
+                     "b": {"band": bands[2],
+                           "min_val": sys.maxsize,
+                           "max_val": -1*sys.maxsize,
+                           "pix_vals": []}
+        }
+    for col in band_dict.keys():
+        im = Image.open(input_filebase+"."+band_dict[col]["band"]+".tif")
+        pix = im.load()
+        ## find the minimum and maximum pixel values in the original scale
+        print("Found image of size {}".format(im.size))
+        for ix in range(im.size[0]):
+            for iy in range(im.size[1]):
+                if pix[ix,iy]> band_dict[col]["max_val"]:
+                    band_dict[col]["max_val"]= pix[ix,iy]
+                elif pix[ix,iy] < band_dict[col]["min_val"]:
+                    band_dict[col]["min_val"] = pix[ix,iy]
+        band_dict[col]["pix_vals"] = pix
+    # Take the overall max of the three bands to be the value to scale down with.
+    print("Max values {} {} {}".format(band_dict["r"]["max_val"],
+                                       band_dict["g"]["max_val"],
+                                       band_dict["b"]["max_val"]))
+
+    overall_max = max((band_dict[col]["max_val"] for col in ["r","g","b"]))
+
+    # create a new image where we will fill RGB pixel values from 0 to 255
+    get_pix_val = lambda ix, iy, col: \
+        max(0, int(band_dict[col]["pix_vals"][ix,iy] * 255/ \
+#                   band_dict[col]["max_val"]
+                   (overall_max+1)
+        ))
+    new_img = Image.new("RGB", im.size)
+    for ix in range(im.size[0]):
+        for iy in range(im.size[1]):
+            new_img.putpixel((ix,iy), tuple(get_pix_val(ix,iy,col) \
+                                            for col in ["r","g","b"]))
+    return new_img
